Feature/stability 2

+++ ZenPacks/zenoss/Import4/bin/prep_perf_import.sh
@@ -1,4 +1,4 @@
-#! /usr/bin/env bash
+#! /bin/bash
 ##############################################################################
 #
 # Copyright (C) Zenoss, Inc. 2014-2015, all rights reserved.
@@ -8,6 +8,7 @@
 #
 ##############################################################################
 
+# executed as root
 export fail_records="/import4/perf.fail.records" # keep the failed status for top level UI
 
 # common block
@@ -22,16 +23,19 @@ tsdb
 "
 
 save_dir="/import4/Q.save"
-rm -rf "$save_dir"
+
+# save the last run info
+if [[ -d "$save_dir" ]] 
+then
+    rm -rf "$save_dir"
+fi
+
 mkdir -p  "$save_dir"
-chmod -R a+w "$save_dir"
 [ -f "$fail_records" ] && mv "$fail_records" "$save_dir"
 
 # the extra one for tsdb
 mkdir -p "/import4/Q.tsdb/.tmp"
 mkdir -p "/import4/Q.tsdb/.fail"
-chmod -R 777 "/import4/Q.tsdb"
-chown -R zenoss:zenoss "/import4/Q.tsdb"
 
 for dname in $targets
 do
@@ -41,10 +45,9 @@ do
     # recreate the struct
     mkdir -p "/import4/Q.$dname/.done"
     chmod -R 777 "/import4/Q.$dname"
-    chown -R zenoss:zenoss "/import4/Q.$dname"
 done
 
 # print out the dir structures
-find /import4 -type d -printf "%M %u %p"
+find /import4 -maxdepth 2 -type d -printf "%M %u %p\n"
 
 info_out "Performance directories cleaned up"
